+++ src/index.ts
@@ -1,8 +1,7 @@
 // Main exports
 export { SarosAMMPair } from './services/pair';
-export { SarosBaseService, MODE, type SarosAMMConfig } from './services/base';
-export { SarosFarm, MODE as FARM_MODE, type SarosFarmConfig, type PoolType } from './services/farm';
-export { SarosStake, MODE as STAKE_MODE, type SarosStakeConfig } from './services/stake';
+export { SarosBaseService, type SarosAMMConfig } from './services/base';
+export { SarosFarm, type SarosFarmConfig, type PoolType } from './services/farm';
 export { SarosAPIService } from './services/api';
 
 // Types
@@ -15,8 +14,7 @@ export * from './utils/pda';
 export { SarosAMMError, SarosAMMErrorCode } from './utils/errors';
 
 // Constants
-export * from './constants/amm_idl';
-export * from './constants/farm_idl';
+export * from './constants';
 
 // Re-export Solana/SPL types for convenience
 export type { PublicKey, Transaction, Connection } from '@solana/web3.js';
